refactor: seed folder
