+++ seed/imageCards.js
@@ -1,0 +1,20 @@
+const ImageCard = require('../server/db/models/imageCard');
+
+const imageCards = [
+    new ImageCard ({
+      image: "https://urlme.me/success/typed_a_url/made_a_meme.jpg",
+  }),
+    new ImageCard({
+      image: "https://i.pinimg.com/474x/9f/90/cb/9f90cbef900aab168e772429d9269646.jpg",
+  }),
+    new ImageCard({
+      image: "https://memegenerator.net/img/instances/58481893.jpg",
+  }),
+    new ImageCard({
+      image: "https://memegenerator.net/img/instances/40167808.jpg",
+  }),
+    new ImageCard({
+      image: "https://3.bp.blogspot.com/-KznttM4xiDI/UvZY7V1JBEI/AAAAAAAAMio/ZU_H2J-MVFQ/s1600/teacher+meme+long+url+use+best+url+shortener.png",
+  })];
+
+  module.exports = imageCards;
