+++ seed/players.js
@@ -1,0 +1,30 @@
+const Player = require('../server/db/models/player');
+
+const players = [
+    new Player({
+        name: "Danielle",
+        score: 4,
+        sentenceCard: "Test2",
+  }),
+    new Player({
+        name: "Maria",
+        score: 2,
+        sentenceCard: "Hi, this is a test",
+    }),
+    new Player({
+        name: "Anna",
+        score: 3,
+        sentenceCard: "Hola!",
+    }),
+    new Player({
+        name: "Samantha",
+        score: 3,
+        sentenceCard: "Testing!",
+    }),
+    new Player({
+        name: "Jen",
+        score: 4,
+        sentenceCard: "Last test!",
+    })];
+
+    module.exports = players;
